Prevent large emoji from being printed on the window's right edge
This is to prevent scrolling. An emoji printed on the edge of the window
gets split in half, the second half on the next line. If the emoji is
printed in the bottom right corner, the window scrolls down so that the
second half will be in view when it appears.